Add server (httpd)
